i10-moved createCategoryScreen into new folder
for organization purposes
